+++ packages/zowe-explorer/src/extension.ts
@@ -72,7 +72,8 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
     // Determine the runtime framework to support special behavior for Theia
     globals.defineGlobals(preferencesTempPath);
 
-    hideTempFolder(getZoweDir());
+    const zoweDir = getZoweDir();
+    hideTempFolder(zoweDir);
 
     try {
         globals.initLogger(context);
@@ -91,8 +92,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         await readConfigFromDisk();
     } catch (err) {
         globals.LOG.error(err);
-        const errorMessage = localize("initialize.profiles.error", "Error reading or initializing Zowe CLI profiles.");
-        vscode.window.showWarningMessage(`${errorMessage}: ${err.message}`);
+        ZoweExplorerExtender.showZoweConfigError(zoweDir);
     }
 
     // Initialize profile manager